[Refactor] 토스트 메시지 보내는 함수 리팩토링
